// src/shared/enums/permission.enum.ts
export enum PermissionEnum {
  USER_CREATE = 'USER_CREATE',
  USER_UPDATE = 'USER_UPDATE',
  USER_UPDATE_OTHER = 'USER_UPDATE_OTHER',
  USER_ACTIVATE = 'USER_ACTIVATE',
  USER_LINK_ROLE = 'USER_LINK_ROLE',
  USER_LINK_PERMISSION = 'USER_LINK_PERMISSION',
  USER_READ = 'USER_READ',

  ROLE_CREATE = 'ROLE_CREATE',
  ROLE_UPDATE = 'ROLE_UPDATE',
  ROLE_LINK_PERMISSION = 'ROLE_LINK_PERMISSION',
  ROLE_READ = 'ROLE_READ',

  PERMISSION_READ = 'PERMISSION_READ',
}